Minor: Made work for logical input image
git-svn-id: https://svn.broadinstitute.org/CellProfiler/trunk/CellProfiler@7586 1fc53939-2000-0410-845c-e8453a809027

+++ Modules/MeasureObjectIntensity.m
@@ -329,7 +329,7 @@ for i = 1:length(ObjectNameList)
         %%% image.
         mask = (LabelMatrixImage > 0);
         masked_labels = LabelMatrixImage(mask);
-        masked_intensity = OrigImage(mask);
+        masked_intensity = double(OrigImage(mask));
         [x,y] = meshgrid(1:size(LabelMatrixImage,1),1:size(LabelMatrixImage,2));
         masked_x = x(mask);
         masked_y = y(mask);
